requirements.txt
adblockparser==0.7
amqp==2.5.2
anyjson==0.3.3
attrs==19.3.0
Automat==0.8.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
decorator==4.4.1
Django==1.11
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-dynamic-scraper==0.13.2
django-timezone-field==3.1
funcparserlib==0.3.6
future==0.17.1
h2==3.1.1
hpack==3.0.0
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
jsonpath-rw==1.4.0
kombu==4.6.6
lxml==4.4.1
more-itertools==7.2.0
parsel==1.5.2
Pillow==5.4.1
ply==3.11
priority==1.3.0
Protego==0.1.15
psutil==5.6.5
psycopg2==2.8.4
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
python-crontab==2.4.0
python-dateutil==2.8.1
pytz==2019.3
qt5reactor==0.5
queuelib==1.5.0
redis==3.3.11
requests==2.22.0
Scrapy==1.5.0
scrapy-djangoitem==1.1.0
scrapy-proxies==0.4
scrapy-splash==0.7.2
scrapyd==1.2.1
scrapyd-client==1.1.0
selenium==3.141.0
service-identity==18.1.0
six==1.12.0
splash==3.4
sqlparse==0.3.0
Twisted==19.7.0
urllib3==1.25.6
vine==1.3.0
w3lib==1.21.0
xvfbwrapper==0.2.9
zipp==0.6.0
zope.interface==4.6.0
